topic/config-cleanup: 	- cleaned up dash
